【Weather】リソースの修正

+++ database/migrations/2021_01_07_152718_create_weather_records_table.php
@@ -18,9 +18,9 @@ class CreateWeatherRecordsTable extends Migration
             $table->date('day')->comment('日付');
             $table->integer('weather')->comment('天候');
             $table->double('temp')->comment('外気温');
-            $table->integer('humidity')->comment('外湿度');
+            $table->double('humidity')->comment('外湿度');
             $table->double('room_temp')->comment('内気温');
-            $table->integer('room_humidity')->comment('内湿度');
+            $table->double('room_humidity')->comment('内湿度');
             $table->timestamp('created_at')->nullable()->comment('登録日時');
             $table->timestamp('updated_at')->nullable()->comment('更新日時');
         });
